refactor: extract shared markdown helper from WorksHelper
