+++ app/helpers/markdown_helper.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+require "redcarpet"
+
+module MarkdownHelper
+  def markdown(text)
+    return "" if text.blank?
+
+    renderer = Redcarpet::Render::HTML.new(
+      hard_wrap: false,
+      link_attributes: { target: "_blank", rel: "noopener" }
+    )
+
+    markdown = Redcarpet::Markdown.new(renderer,
+      autolink: true,
+      tables: true,
+      fenced_code_blocks: true,
+      strikethrough: true,
+      superscript: true,
+      space_after_headers: true
+    )
+
+    markdown.render(text).html_safe
+  end
+end
